Simulated Annealing Discrete Case
J is the accuracy function. Please return error = 1-accuracy for each 3-tuple requested by this function.

range of epsilon is [0,8]
range of alpha is [0,50]
range of dropout is [0,9]

J is a 3-dimensional discrete cubic simulating training accuracy function. We won't need to floor the indicies once we have training accuracy returned for each epsilon, alpha, dropout.